auto_post/requirements.txt
anthropic>=0.25.0
pytrends>=4.9.0
tweepy>=4.14.0
selenium>=4.18.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.28.0
